refactor(core): 状态契约上移至共享内核叶 e2m2e/status.py，integrators 解除对 algorithm/data 的依赖 (#545)
ConvergenceState/FailureCause（原 data/templates/enums.py）与
CAUSE_STATUS/ResultStatus（原 algorithm/results.py）唯一归属包根
e2m2e/status——零内部依赖的共享内核叶（仿 exceptions.py 先例）。
原路径原位 re-export 保持对象身份，85+17 处既有 import 零改动。
integrators.py 门面不再上溯 algorithm/data（ADR 0012 修订案的第一条缝，
ADR 0039 落地后由检查器 v2 执法）。
